+++ itk-center/itk-dao/src/main/java/com/itk/payment/mapper/RefundFlowMapper.java
@@ -4,22 +4,21 @@ import com.itk.payment.model.RefundFlow;
 import com.itk.payment.model.RefundFlowExample;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
-import org.apache.ibatis.annotations.Select;
 
 public interface RefundFlowMapper {
     int countByExample(RefundFlowExample example);
 
     int deleteByExample(RefundFlowExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(RefundFlow record);
 
-    int insertSelective(RefundFlow record);
+    Long insertSelective(RefundFlow record);
 
     List<RefundFlow> selectByExample(RefundFlowExample example);
 
-    RefundFlow selectByPrimaryKey(Integer id);
+    RefundFlow selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") RefundFlow record, @Param("example") RefundFlowExample example);
 
@@ -28,5 +27,4 @@ public interface RefundFlowMapper {
     int updateByPrimaryKeySelective(RefundFlow record);
 
     int updateByPrimaryKey(RefundFlow record);
-
 }
